backend/requirements.txt
Django==5.0
django-timezone-field<=6.1.0,<6.2
django-celery-beat<=2.6.0,<2.7
celery>=5.4.0,<5.5
redis>=5.0.5,<5.1
psycopg2>=2.9.9,<3.0
django-environ>=0.11.2,<0.12
djangorestframework>=3.15.1,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4
drf-spectacular>=0.27.2,<0.28
django-redis>=5.4.0,<5.5
stripe>=9.12.0,<9.13
django-cors-headers>=4.3.1,<4.4
coverage>=7.5.3,<7.6
requests>=2.31.0,<2.32
pandas>=2.2.1,<2.3
joblib==1.3.2
scikit-learn==1.4.1.post1
xgboost>=2.0.3,<2.1